+++ classes/event/invitation_attempted.php
@@ -51,9 +51,10 @@ class invitation_attempted extends invitation_base {
     }
 
     public function get_description() {
-        $description = "The user with id {$this->userid} attempted to accept an invitation " .
-                "for course with id '{$this->other['courseid']}'";
-        $description .= property_exists((object)$this->other, 'errormsg') ? " with error '{$this->other['errormsg']}" : '';
+        $userid = empty($this->userid) ? get_string('anonymoususer', 'enrol_invitation') : $this->userid;
+        $errormsg = property_exists((object)$this->other, 'errormsg') ? $this->other['errormsg'] : '';
+        $description = get_string('failuredescription', 'enrol_invitation',
+                ['userid' => $userid, 'courseid' => $this->other['courseid'], 'errormsg' => $errormsg]);
         return $description;
     }