+++ classes/event/invitation_rejected.php
@@ -48,12 +48,20 @@ class invitation_rejected extends invitation_base {
     }
 
     public static function get_name() {
-        return get_string('event_invitation_accepted', 'enrol_invitation');
+        return get_string('event_invitation_rejected', 'enrol_invitation');
     }
 
     public function get_description() {
-        $description = "The user with id {$this->userid} rejected an invitation for course with id '{$this->other['courseid']}'";
-        $description .= property_exists((object)$this->other, 'errormsg') ? ' and wasn\'t logged in.' : '';
+        $userid = empty($this->userid) ? get_string('anonymoususer', 'enrol_invitation') : $this->userid;
+        if (property_exists((object)$this->other, 'errormsg')) {
+            // Failure.
+            $description = get_string('failuredescription', 'enrol_invitation',
+                    ['userid' => $userid, 'courseid' => $this->other['courseid'], 'errormsg' => $this->other['errormsg']]);
+        } else {
+            // Success.
+            $description = get_string('rejecteddescription', 'enrol_invitation',
+                    ['userid' => $userid, 'courseid' => $this->other['courseid']]);
+        }
         return $description;
     }