+++ invitation_form.php
@@ -42,7 +42,7 @@ class invitation_form extends moodleform {
      * The form definition.
      */
     public function definition() {
-        global $CFG, $DB, $USER;
+        global $CFG, $USER, $COURSE;
         $mform = & $this->_form;
 
         // Get rid of "Collapse all" in Moodle 2.5+.
@@ -51,7 +51,6 @@ class invitation_form extends moodleform {
         }
 
         // Add some hidden fields.
-        $course = $this->_customdata['course'];
         $course = $this->_customdata['course'];
         $instance = $this->_customdata['instance'];
         $context = $this->_customdata['context'];
@@ -340,7 +339,7 @@ class invitation_email_form extends moodleform {
      * The form definition.
      */
     public function definition() {
-        global $CFG, $DB, $USER;
+        global $CFG, $USER;
         $mform = & $this->_form;
         // Get rid of "Collapse all" in Moodle 2.5+.
         if (method_exists($mform, 'setDisableShortforms')) {
@@ -350,6 +349,7 @@ class invitation_email_form extends moodleform {
         $course = $this->_customdata['course'];
         $instance = $this->_customdata['instance'];
         $context = $this->_customdata['context'];
+        $prefilled = $this->_customdata['prefilled'];
         $mform->addElement('hidden', 'courseid');
         $mform->setType('courseid', PARAM_INT);
         $mform->setDefault('courseid', $course->id);
@@ -411,6 +411,16 @@ class invitation_email_form extends moodleform {
         $mform->addHelpButton('message', 'emailmsghtml', 'enrol_invitation',
         get_string('message_help_link', 'enrol_invitation'));
         // Check for correct email formating later in validation() function.
+
+        // Set defaults if the user is resending an invite that expired.
+        if (!empty($prefilled)) {
+            $mform->setDefault('role_group[roleid]', $prefilled['roleid']);
+            $mform->setDefault('email', $prefilled['email']);
+            $mform->setDefault('subject', $prefilled['subject']);
+            $mform->setDefault('message', $prefilled['message']);
+            $mform->setDefault('show_from_email', $prefilled['show_from_email']);
+            $mform->setDefault('notify_inviter', $prefilled['notify_inviter']);
+        }
         $this->add_action_buttons(false, get_string('inviteusers', 'enrol_invitation'));
     }