+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [1.3.1] - 2022-01-13 (DEV-BETA)
+## [1.3.2] - 2022-01-18 (DEV-BETA)
 ### Added
 - Initial public re-release on Moodle.org and GitHub by TNG Consulting Inc.
 - Fix-8: Default email subject is now customizable.
 - Fix-2: Invitation text is now editable in language customization editor.
+- You can now re-send an invite if you accidentally revoke an invitation.
 - Ability to customize the default email subject line.
 - Many new variables now available in the email template.
 - Anonymous users (no account) can reject invitation to users without an account. Users with an account will need to login.
@@ -15,15 +16,20 @@ All notable changes to this project will be documented in this file.
 - CHANGELOG.md.
 - composer.json
 ### Updated
+- Fix-19: The Moodle log entries from this plugin are now displayed using strings from the language pack.
+- Fix-18: No longer displays Undefined Variable COURSE when using invitations with default values set to No.
+- Fix-17: Invitation Acceptance and Rejection failures are now logged.
 - Fix-16: Email footer will now include primary admin's email address if support email address is not defined.
 - Fix-12: Email FROM is now set to primary admin's id if show_from_email is disabled and support email address is not defined.
 - Fix-14: Help bubble on email form now shows correct preview of email message body.
 - Fix-13: Email footer now supports multiple languages.
+- Fix-11: Resend Invite link now prefills invitation form.
 - Fix-10: Extend invite link option now works.
 - Fix-9: Email address and course name fields are now wider when default values are disabled.
+- Fix-7: The default message is now displayed when sending new invitations
 - Fixed multi-language issues.
 - Improved default email message. Note: You may need to update language strings in other languages.
-- Updated .gitignore
+- Updated .gitignore.
 - Multiple fixes for Multi-Language Moodle environments.
 - Improved accessibility and rtl language of invitation emails.
 - Plugin now compatible and tested with Moodle 3.9, 3.10 and 3.11.